+++ test/open_pgp/public_key_encrypted_session_key_packet_test.exs
@@ -1,5 +1,6 @@
 defmodule OpenPGP.PublicKeyEncryptedSessionKeyPacketTest do
   use OpenPGP.Test.Case, async: true
+  doctest OpenPGP.PublicKeyEncryptedSessionKeyPacket
 
   alias OpenPGP.Packet
   alias OpenPGP.Packet.PacketTag
@@ -33,6 +34,17 @@ defmodule OpenPGP.PublicKeyEncryptedSessionKeyPacketTest do
     end
   end
 
+  describe ".encode/3" do
+    test "encodes packet body" do
+      ciphertext = "Ciphertext"
+      public_key_id = "6BAF2C48"
+      public_key_algo = {16, "Elgamal (Encrypt-Only) [ELGAMAL] [HAC]"}
+
+      assert body = PublicKeyEncryptedSessionKeyPacket.encode(ciphertext, public_key_id, public_key_algo)
+      assert <<3::8, public_key_id::binary, 16::8, ciphertext::binary>> == body
+    end
+  end
+
   describe ".decrypt/2" do
     test "decrypts key material given a valid decrypted Secret-Key Packet" do
       [
@@ -61,6 +73,56 @@ defmodule OpenPGP.PublicKeyEncryptedSessionKeyPacketTest do
       assert {m_e_mod_n} = session_key_material
 
       assert "26A582ACA833B8EE60CC58865D19A21653D38CB0737125C9ABF973405E3B233C" = Base.encode16(m_e_mod_n)
+    end
+  end
+
+  describe ".encrypt/2" do
+    # [RFC3526](https://datatracker.ietf.org/doc/html/rfc3526)
+    @modp_group_1536 """
+    FFFFFFFF FFFFFFFF C90FDAA2 2168C234 C4C6628B 80DC1CD1
+    29024E08 8A67CC74 020BBEA6 3B139B22 514A0879 8E3404DD
+    EF9519B3 CD3A431B 302B0A6D F25F1437 4FE1356D 6D51C245
+    E485B576 625E7EC6 F44C42E9 A637ED6B 0BFF5CB6 F406B7ED
+    EE386BFB 5A899FA5 AE9F2411 7C4B1FE6 49286651 ECE45B3D
+    C2007CB8 A163BF05 98DA4836 1C55D39A 69163FA8 FD24CF5F
+    83655D23 DCA3AD96 1C62F356 208552BB 9ED52907 7096966D
+    670C354E 4ABC9804 F1746C08 CA237327 FFFFFFFF FFFFFFFF
+    """
+    @prime_p @modp_group_1536 |> String.replace(~r/[^0-9ABCDEF]/, "") |> Base.decode16!()
+    @group_g <<2::8>>
+    test "encrypts AES-256 session key with Elgamal" do
+      alias PublicKeyEncryptedSessionKeyPacket, as: PKESK
+
+      # Define Diffie-Hellman (DH) parameters (p and g). These are commonly used predefined values.
+      p = :binary.decode_unsigned(@prime_p)
+      g = :binary.decode_unsigned(@group_g)
+
+      # Generate private key such as "1 < private_key < p-1"
+      # Any 1024-bit (128 bytes) big-endian will be smaller than 1536-bit big-endian.
+      private_key = :crypto.strong_rand_bytes(128)
+      a = :binary.decode_unsigned(private_key)
+
+      # Generate the public key (g**private_key mod p)
+      e = :crypto.mod_pow(g, a, p)
+
+      public_key_material = {@prime_p, @group_g, e}
+      public_key_algo = {16, "Elgamal (Encrypt-Only) [ELGAMAL] [HAC]"}
+
+      session_key_algo = {9, "AES with 256-bit key"}
+      session_key = "12345678901234567890123456789012"
+
+      assert ciphertext = PKESK.encrypt(session_key, session_key_algo, public_key_material, public_key_algo)
+
+      # Decrypt Elgamal
+      assert {c1, next} = Util.decode_mpi(ciphertext)
+      assert {c2, <<>>} = Util.decode_mpi(next)
+
+      x = :crypto.mod_pow(c1, a, p)
+      y = :crypto.mod_pow(x, p - 2, p)
+      decoded_value = rem(:binary.decode_unsigned(c2) * :binary.decode_unsigned(y), p)
+      plaintext = :binary.encode_unsigned(decoded_value)
+
+      assert [_, <<0x09, "12345678901234567890123456789012", _::16>>] = String.split(plaintext, <<0>>, parts: 2)
     end
   end
 end